feat(deepresearch): refactor InfoCheckTool to InfoCheckService && add MAX/MIN OptimizeQueryNum
